fix: BIGINT farm_id, 30s delay between farm snapshots
- ALTER farm_id INTEGER -> BIGINT for large farm IDs
- Add 30s sleep between farms to avoid SFL API rate limits
- Farm 155498 always processed first

+++ azure-functions/schema.sql
@@ -4,7 +4,7 @@
 -- Farm snapshots: full JSON + computed diff
 CREATE TABLE farm_snapshots (
   id BIGSERIAL PRIMARY KEY,
-  farm_id INTEGER NOT NULL,
+  farm_id BIGINT NOT NULL,
   captured_at TIMESTAMPTZ NOT NULL DEFAULT NOW(),
   game_data JSONB NOT NULL,
   diff JSONB,
